requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1
